// frontend/bindings/collections/internal/model/models.ts
export class Attachment {
    "filename": string;
    "label"?: string;

    /** Creates a new Attachment instance. */
    constructor($$source: Partial<Attachment> = {}) {
        if (!("filename" in $$source)) {
            this["filename"] = "";
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new Attachment instance from a string or object.
     */
    static createFrom($$source: any = {}): Attachment {
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        return new Attachment($$parsedSource as Partial<Attachment>);
    }
}
// ...
export class Site {
    "id": string;
    "title": string;
    "domain": string;
    "icon"?: string;
    "cover"?: string;
    "description"?: string;
    "tags": string[];
    "attachments": Attachment[];
    "bookmark_count": number;
    "created_at": time$0.Time;
    "updated_at": time$0.Time;

    /** Creates a new Site instance. */
    constructor($$source: Partial<Site> = {}) {
        if (!("id" in $$source)) {
            this["id"] = "";
        }
        if (!("title" in $$source)) {
            this["title"] = "";
        }
        if (!("domain" in $$source)) {
            this["domain"] = "";
        }
        if (!("tags" in $$source)) {
            this["tags"] = [];
        }
        if (!("attachments" in $$source)) {
            this["attachments"] = [];
        }
        if (!("bookmark_count" in $$source)) {
            this["bookmark_count"] = 0;
        }
        if (!("created_at" in $$source)) {
            this["created_at"] = null;
        }
        if (!("updated_at" in $$source)) {
            this["updated_at"] = null;
        }

        Object.assign(this, $$source);
    }

    /**
     * Creates a new Site instance from a string or object.
     */
    static createFrom($$source: any = {}): Site {
        const $$createField6_0 = $$createType0;
        const $$createField7_0 = $$createType2;
        let $$parsedSource = typeof $$source === 'string' ? JSON.parse($$source) : $$source;
        if ("tags" in $$parsedSource) {
            $$parsedSource["tags"] = $$createField6_0($$parsedSource["tags"]);
        }
        if ("attachments" in $$parsedSource) {
            $$parsedSource["attachments"] = $$createField7_0($$parsedSource["attachments"]);
        }
        return new Site($$parsedSource as Partial<Site>);
    }
}
// ...
const $$createType0 = $Create.Array($Create.Any);
const $$createType1 = Attachment.createFrom;
const $$createType2 = $Create.Array($$createType1);